Minor refactoring, added new Uri extension
